python/dlms_json/requirements.txt
gurux-common==1.0.13
gurux-dlms==1.0.180
gurux-net==1.0.22
gurux-serial==1.0.21
PyMySQL==1.1.0
pymysql-pooling==1.0.6